feat: add fancy_logger db export test
